backend/requirements.txt
anyio==3.6.1
autopep8==1.6.0
click==8.1.3
fastapi==0.78.0
h11==0.13.0
idna==3.3
markdown2==2.4.3
pycodestyle==2.8.0
pydantic==1.9.1
python-frontmatter==1.0.0
PyYAML==6.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
typing-extensions==4.3.0
uvicorn==0.18.2
